requirements.txt
pyenchant==3.2.0
PySimpleGUI==4.43.0
PyDictionary==2.0.1